% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{CommitID}
\alias{CommitID}
\title{Get Commit ID of the Source SymEngine Library}
\usage{
CommitID()
}
\value{
A length-one character vector.
}
\description{
The commit id of the source repository of SymEngine is specified in
the DESCRIPTION file. This function simply reads it out.
}
