% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{PkgLibsFlags}
\alias{PkgLibsFlags}
\title{Get Compilation Flags for Linking}
\usage{
PkgLibsFlags()
}
\value{
It prints and invisibly returns the list of flags as a character vector.
}
\description{
Util function used in "symengine" R package.
}
